requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
instagrapi==1.4.6
Pillow==8.1.0
praw==7.1.0
prawcore==1.5.0
pydantic==1.7.3
PySocks==1.7.1
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0